Avoid creating TrainerConfig during parse_conf_file() (#774)
* Avoid creating TrainerConfig during parse_conf_file()

Before this change, get_observation() will create a TrainerConfig(). This makes the configurable
values of TrainerConfig to be all labelled as used. And it will not be allowed to change after get_observation().

Now we use get_config() to get the value of TrainerConfig.data_transformer_ctor
to avoid creating TrainerConfig object. So the configs of other TrainerConfig values
can still be changed after get_observation_spec().

Also changed sac_carla.gin to use alf-config

* parse_conf_file always create global env and init seed

* Better handling of --conf_param

Previously, if the related module has not been imported, the config
name specified through --conf_param will not be found. This make it
very hard to use since the related modules are typically imported
in the config file. This change delay the binding of conf_param
until the related module has been loaded.

* Update comments of sac_carla_conf.py